refactor: move header in header actions file
